Criado e configurado exception de Produto; Criado e configurado controller, entity, repository e service de Usuario

+++ src/main/java/com/th_oliveira/spring_boot_project/service/ProdutoService.java
@@ -1,12 +1,11 @@
 package com.th_oliveira.spring_boot_project.service;
 
 import com.th_oliveira.spring_boot_project.entity.ProdutoEntity;
+import com.th_oliveira.spring_boot_project.exceptions.RecursoNaoEncontradoException;
 import com.th_oliveira.spring_boot_project.repository.ProdutoRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProdutoService {
@@ -21,8 +20,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<ProdutoEntity> buscarPorId(Long id){
-        return produtoRepository.findById(id);
+    public ProdutoEntity buscarPorId(Long id){
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Produto com ID " + id + " não encontrado!"));
     }
 
     public ProdutoEntity salvarProduto(ProdutoEntity produtoEntity){
@@ -30,6 +30,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id){
+        if(!produtoRepository.existsById(id)){
+            throw new RecursoNaoEncontradoException("Produto com ID " + id + " não encontrado!");
+        }
         produtoRepository.deleteById(id);
     }
 }
